Před trailerem: přetečení populace, tempo guvernéra a ulice v hromadné stavbě
POPULACE −2,15 MLD. Kapacita bydlení se držela v int, jenže násobič z Vzestupu
se skládá násobně přes desítky úrovní: jediný vylepšený dům přidá řádově víc,
než se do 32 bitů vejde. Kapacita přetekla do záporu a protože je stropem
růstu, stáhla populaci s sebou. Teď je to double; metriky, rekordy a snímky
historie procházejí přes Numbers.ToLong, který ořízne na kraj rozsahu místo
aby překlopil do záporu. Strop růstu se navíc nikdy nesmí dostat pod nulu.

TEMPO AUTO-STAVBY. Rozpočet staveb na interval se počítal přes
(int)Math.Round(…) uprostřed výpočtu — na maximu vyšlo číslo mimo rozsah int,
přeteklo a Math.Max(1, …) ho srazil na JEDNU stavbu za interval. Čím víc měl
hráč vylepšení, tím pomaleji se stavělo. Počítá se v double a ořezává až na
konci (strop 2048 za interval kvůli snímkové frekvenci, ne kvůli balancu).
Nad jednu stavbu se dláždí až po dávce, stejně jako u hromadné stavby hráče.

TEMPO VYLEPŠOVÁNÍ. Stupeň guvernéra říká, CO se smí vylepšovat; kolik toho za
interval stihne, řídí totéž tempo Vzestupu. Dřív to byly tři budovy za interval
bez ohledu na cokoli, což u města o tisících domů znamenalo čekat hodiny.
Hledání navíc nezačíná od nuly, ale od místa odvozeného z tiku — tik je
v savu, takže načtená hra pokračuje úplně stejně jako ta původní.

ULICE V HROMADNÉ STAVBĚ. Tažení stavělo souvislý slitek bez jediné volné
dlaždice, takže se auto-silnice neměla kudy dostat dovnitř a mohla čtvrť jen
obkroužit zvenčí. Velký tah teď nechává pruhy pro ulice; malý ne, protože kdo
táhne řadu tří chalup, chce tři chalupy. U větších půdorysů se hlídá celý
obrys, ne jen roh.

GUVERNÉR NESTAVĚL VÝROBU. Strop pracovních míst byl tvrdý zákaz: jakmile
počet míst přerostl populaci × 2,5, přestal stavět cokoli s dělníky, a to
natrvalo — zbyly mu domy (nula míst) a pole (hlad kontrolu obchází). Teď platí
jen tehdy, když už někde stojí budova bez lidí. Plná zaměstnanost není důvod
nestavět; když má město lidi navíc, je další výrobna přesně to, co potřebuje.

Co jsem NEPŘIDAL: zkoušel jsem k tomu ještě proaktivní potřebu „dochází
stavební materiál", aby guvernér stavěl doly dřív, než mu něco dojde. Rozešla
determinismus mezi běžící a načtenou hrou (stejné budovy, stejné potřeby, jiná
zásoba) a příčinu jsem nedohledal, takže jsem ji zahodil celou. Nahlášený
příznak řeší už uvolněný strop pracovních míst.

+++ tests/CivDle.Core.Tests/Sim/DemolishMoveTests.cs
@@ -24,7 +24,7 @@ public class DemolishMoveTests
         var (content, sim) = Grassland();
         int house = content.Buildings.IndexOf("house");
         int wood = content.Resources.IndexOf("wood");
-        int capBefore = sim.HousingCapacity;
+        double capBefore = sim.HousingCapacity;
 
         Assert.Equal(PlacementResult.Ok, sim.TryPlaceBuilding(house, 2, 2));
         Assert.True(sim.IsOccupied(2, 2));
